[Feature #160947820] Separate test feedback
