+++ server/tests/controllers/Authorization.test.js
@@ -1,0 +1,45 @@
+import chai from 'chai';
+import chaiHttp from 'chai-http';
+import app from '../../app';
+
+
+chai.use(chaiHttp);
+const { expect } = chai;
+
+
+describe('AUTHORIZE USER TOKEN /', () => {
+  const name = 'somename';
+  const description = 'a description message here';
+  const price = '700';
+  const imageurl = 'http://sometesturl';
+
+  it('Acess denied. Invalid token', async () => {
+    try {
+      const res = await chai.request(app)
+        .post('/api/v1/menu')
+        .set('x-auth-token', 'xxxxxxxxxxxxxxxx')
+        .send({
+          name, description, price, imageurl
+        });
+      expect(res.status).to.equal(400);
+      expect(res.body).to.be.an('object');
+    } catch (err) {
+      throw err.message;
+    }
+  });
+
+  it('Acess denied. No token provided', async () => {
+    try {
+      const res = await chai.request(app)
+        .post('/api/v1/menu')
+        .set('x-auth-token', '')
+        .send({
+          name, description, price, imageurl
+        });
+      expect(res.status).to.equal(401);
+      expect(res.body).to.be.an('object');
+    } catch (err) {
+      throw err.message;
+    }
+  });
+});
